+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [(GH-341)](https://github.com/puppetlabs/pdkgo/issues/341) `prm build` command to package tools for distribution
+
 ## [0.1.0]
 
 ### Added